01 - Simple Chatbot/pyproject.toml
[project]
name = "session01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=5.45.0",
    "openai>=1.107.1",
    "python-dotenv>=1.1.1",
]
